Implement database block and record structure
This commit implements the block and record structure for the database.
A block is a contiguous chunk of memory that holds zero or more
records. A record is a single key value pair.

The block layout takes some inspiration from PostgreSQL [0]. The record
layout and diagrams are inspired by the mini-LSM project [1]. The code
also takes some inspiration from Database Design and Implementation
[2].

[0] https://www.postgresql.org/docs/17/storage-page-layout.html
[1] https://skyzh.github.io/mini-lsm/week1-03-block.html
[2] E. Siore, Design and implementation of database management systems (M.: DMK Press, 2020)